Altered macro syntax into a function-call-style syntax.
This revision replaces <<macroname arg arg>> macro tags with macroname(arg arg) macro syntax. The motivation for this is to make a step to uniting the concepts of functions and macros, two similar but disparite entities, and to eliminate the somewhat clunky double-chevron SGML-esque tags of TiddlyWiki. More thoughts on this, especially regarding the readability of this new syntax, are [here](https://twitter.com/webbedspace/status/479816144165945344).

Also:

 * Made Macros.create() rely on TwineMarked.RegExpStrings to decompose the macro's call string into its components.

 * Added Macros.instantiate() as a factory function that reliably creates HookMacroInstances or MacroInstances depending on the macro. (Previously, ternaries were needed at each callsite.)

 * Added some RegExpStrings for Twine code entities (number, string, boolean). These are now tokenised by TweeLexer, but Renderer doesn't treat them differently to text.

 * Altered TwineMarked's lexer to parse recursive macro syntax. This change is fairly clunky, and will be replaced soon, possibly by the XRegExp lib.

 * Altered Debug Mode's CSS to highlight <tw-align> spans.