Macro arguments are now stored as thunks.
This lays the groundwork for 'live' macros, whose arguments are stored and lazily evaluated. For instance, I'm thinking of adding a "when($a > 2)" style macro, and it should be able to constantly re-evaluate its argument in the future, not just at the call site.

At the moment, the macros' implementations have not been substantially altered, and the thunk is simply unwrapped immediately on call, producing no net change.

Also:

 * Partially implemented the idea of a "changer macro", a macro that simply performs a transformation on its given hook, and which has a particular protocol ("changer functions" which mutate "ChangerDescriptor" objects) facilitating composition with other changer macros. At the moment, only nobr() is altered to use this functionality.

 * Changed the call signature of Engine.render(), and caused myself a lot of bother.

 * Temporarily(?) disabled single-quotes for TwineScript string literals, due to perf problems with the string regex.

 * Removed some residual MacroInstance references.

 * Now, macro arguments that evaluate to Error objects are properly accounted for and returned instead of executing the macro's function.