Moved some Engine functionality into a new PassageInstance class.
You may have noticed prior to this that Engine, MacroLib and so forth have a problem whereby a jQuery "top" variable, representing the DOM nodes visible to the macros/code, must be passed ceaselessly down through their various calls. This commit doesn't quite alleviate the problem, but it makes the "top" variable more extensible and intelligible.

Now, the "top" is a PassageInstance, an object with a "dom" property holding the previous jQuery collecion, various DOM visibility-sensitive Engine methods (render, refreshEnchantments) as own methods, and  the potential to be augmented with more properties to provide greater information to scripts. In fact, one such property ("timestamp") is added, as a means to implement a "time" keyword in TwineScript.

Also:

  * TwineScript: each environ now has its own Operation object, allowing keywords like "time" to be uniquely bound to it.

  * Added Utils.assert(), a rudimentary assertion checker, to accompany Utils.impossible().