Made hook tag syntax bi-directional.
I want to try this for a moment. As you know, the link syntax is bi-directional insofar as passage name and link text are concerned - you can write [[passage name<-link text]] or [[link text->passage name]]. What if hook tags were asymmetrical - <tag| and |tag> instead of [tag] - a gift tag shape? Now, you can place the tag at the start of the hook instead of the end, if you wish, which may be desirable when the tag has a descriptive, semantic name, serving as a "title" for the hook. Plus, it eliminates that pesky ambiguity I fixed a short while earlier.