Implemented "anonymous hook after macro" syntax.
Now, hooks that appear immediately after macros no longer need tags - the macro's presence permits them to be interpreted as hooks rather than prose. So, the hook-as-block syntax I espoused earlier is now basically implemented. This is a weensy bit shaky, syntax-wise, but I can live with this for at least a little while.

Also:

 * Fixed some bugs in the previous commit.

 * Added String#contains, String#startsWith and String#endsWith to the ES6 shims file.

 * Wrote more comments, as usual.