Merged WordArray into Section, and refactored Section.
It had become clear that WordArray is an artifact of an earlier design period, where the line between TwineScript and Javascript was thinner, and WordArray needed to be a class with a wide API.

This is no longer the case - now, all of WordArray's chief responsibilities, aggregating <tw-hook> elements and pseudo-hooks, is now handled by two very thin classes within Section: HookSet and PseudoHookSet. They only have one method, a forEach iterator, and moreover are not expected to be used outside of Section very much at all.

Also:

 * Moved a number of hook-specific utilities from Utils into HookUtils, a module which currently only Section uses.

 * Disabled the enchantment macros, pending re-implementation as side-effect sensor macros.

 * Updated the .hgignore to ignore the build folder, and .bat files.