Fixed some bugs in the previous commit.
Also added references to a proposed 'gets' operator for TwineScript, to be used in the when() macro as a replacement for the click() macro - you could say "when(?x gets clicked)" (with 'gets' being the operator and 'clicked' being a keyword for an event type). I have my reservations about it, though, and will have to think about it more.