Re-instated enchantment macros.
Enchantment macros are now re-implemented as special kinds of changer macros, which simplifies their implementation in crucial ways.

The single enchanters (click, mouseover, mouseout) are back in, and while the "combos" (click-replace) aren't in yet, they will be in a subsequent commit.

This also eliminates the "enchantment registry" that was in Macros, instead opting for a simpler approach contained entirely in MacroLib.