Fixed complicated TwineMarkup bug where nested syntax (hooks) wasn't working.
You may notice this bugfix involves a major re-working of TwineMarkup's internals. RecursiveExpressions are out, because it was too hard to allow them to nest different kinds of syntax with similar terminator tokens.

Now, there are two passes in lexing: the first pass handles all non-nestable syntax (e.g. bold, links etc.) and, when given the opening or closing symbol for a nestable syntax, simply creates a temporary token. The second pass gathers these temporary tokens and matches them up, folding them and their contained children into a proper token (or demotes it to a text token when a match couldn't be made).

This isn't a fully satisfactory solution, though: since inside-macro-arguments and outside-macro-arguments have different allowed syntax (e.g. * means "multiply" inside macro arguments and "emphasis" outside) then when a macro instance token is folded up, its children must be thrown out and re-lexed. This has a cost, but since macro arguments tend to be fairly small and uncommon, it is not a particularly detrimental cost.