Changed macro syntax again, from macroname(...) to (macroname: ...).
The reasons for this late change are several: firstly, the inability to properly separate the macro name from directly adjacent prose was the main culprit - I had intended some special separator be added for these particular cases, but I relented to reason. Also, the closer similarity of these to Twine 1's syntax, which also placed the macro name inside matching delimiters, was considered.

Despite these changes, the parameter syntax remains as it is: the params are separated by commas, and, additionally, the macro name is separated from them with a colon. This may very slightly inhibit the readability of certain plain-language-y macros like (set: ...), but the incumbent syntax already incurred a similar tax with the placement of an open-bracket.

Also and additioanlly:

 * Fixed some Debug Mode CSS that caused line heights to change due to the presence of a link or macro instance.

 * Fixed hookRefs in passage text (that's things like "?book" as literals in the prose) not displaying the text of the named hook when possible. This fulfills their basic functionality.

 * Updated hooktest.html, for whatever that's worth.