Refactored Lexer and Markup a fair bit.
Now, Markup rules are more explicitly divided into "modes", which are arrays listing the rules that are valid within a particular token. Currently, the modes are "flatMode" and "macroMode" (alongside two currently defunct modes, "inlineMode" and "blockMode").

Also, I've generally improved the setup code for Markup in a number of uninteresting ways.