Partly implemented proper assignment in TwineScript.
Assignments are parsed by TwineScript and converted into an "assignmentRequest" object, which is passed to the (set:) macro for execution. This allows some error-reporting code to be centralised in that macro's implementation, and also allows the act of changing variable state to be proxied at that location, if the need would so be.

Currently, only the (set:) macro may contain an assignment, thus saving on confusion when unintended assignments are made in other macros. Also, assignment currently only permits single variables ($var) or hooks (?hook) as the lvalue. In case you weren't aware, assigning to a hookref changes the text of ALL hooks which match that ref. Cool! More work on other assignment forms (such as destructuring or indexing) is forthcoming.

Also, I partly implemented basic augmented assigment ($var += "x") but still need to work on it.

Also, a number of tiny bugs were fixed in TwineMarkup.