Fixed the (remove:) macro.
It's admittedly a bit unintuitive that (remove:) is a changer macro, given that it doesn't use an attached hook.

Also: added a jQuery plugins module, for storing some convenience methods - popAttr(), which removes an attribute and returns its value, and tag(), which returns the lowercase tag name of the first matched element.