Overloaded - to remove all of an element from arrays and strings.
Since TwineScript currently lacks an adequate "delete" operation, and with a lot of thought, I've made subtraction the operation to remove array elements/string characters. This may seem bizarre, but for what it's worth, Python overloads - to find the difference of sets, though they are admittedly unordered and unique, unlike arrays.

Note that you can only subtract an array from an array, at present - subtracting 1 element from an array requires it be wrapped in an (a:) wrapper.

Also: I've wrapped several other operations (multiply, divide, modulus) so that they will fail loudly if given non-numbers to work with.