Split Macrolib into several smaller files.
The file had gotten a tad unwieldy to navigate. Now, built-in macros are categorised as "values" (value macros), "sensors" (sensor macros), "styleChangers" (changers concerned with CSS) and "enchantments". One issue with the latter is that revision macros like (replace:) must be included with the enchantments as well, due to combos such as (click-replace:) being computed there.

Also: altered the grunt-sass version in package.json to fix a build bug in OS X.