Made ChangeDescriptors into a proper class.
ChangeDescriptors (the 'r' is dropped to reflect the fact that they represent changes, not changers) now inherit from a prototype. The render() private method of Section has been made a public method of this prototype, owing to the fact that it was always strongly tied to the passed ChangeDescriptor anyway.

As a result of this, there should be a slight optimisation: for multiple-instance render targets (like same-named hooks), updateEnchantments() is no longer called until they have all been executed.

Also:

 * Added the repl.js file which was missing in the previous commit.

 * Changed some requirejs references to include the directory, to make Harlowe's file hierarchy a bit clearer.