Partially implemented a (link:) macro.
The point of this macro is to A) provide a more convenient means of immediately making single links that do things, rather than writing a hook and the (click:) macro, and B) to eventually host passage links, via the link syntax desugaring to a (link-goto:) call. This would potentially allow transitions, (set:) macros, etc. to be more easily attached to links, in a less bespoke fashion than in Twine 1. But, it's not quite done yet.

Also:

 * Remembered to update the Gruntfile with the recent folder hierarchy changes.

 * Refactored renderError() within Section, thus allowing sensor macros to generate proper error HTML.

 * Fixed a bug where onlyNumbers() in Operations wouldn't properly propagate errors passed into it.