Added preliminary whitespace-collapsing markup braces.
The { and } braces now denote passage prose whose whitespace must be collapsed - runs of \s and \n are converted to a single space, thus eliminating whitespace.

This implementation has two deficiencies so far: the syntax does not nest, and it affects string literals inside macro tags. Both of these need to be addressed at some point.

Also: renamed the internal ChangeDescriptor 'cloaked' property to 'enabled', and inverted it.