Formally implemented (link-goto:).
(link-goto:) is a command macro that produces a passage link, essentially an eager-argument version of (link:?)[(goto:?)]. The future goal is for the link syntax to desugar into this macro, as I mentioned earlier, and also to allow (transition:) commands to be attached to this. But, at the moment, it acts as a convenient link macro by itself.

Also: altered Colour.create so that it can now accept a CSS3 colour name, and correctly use jQuery#css to convert it to an RGB object. However, this is currently not used anywhere, so it may prove to be extraneous.