Added ES6 spread syntax to macro calls.
By putting the ellipsis "..." operator before an array or string, the elements are changed to positional elements in the macro call, separated by  commas. Hence, (print: ...(a: 2, 4)) is the same as (print: 2, 4). It's standard JavaScript ES6, but it's implemented in ES5 for maximum browser support. I considered briefly the "*" splat symbol used by Ruby and Python, but decided that it clashed with the variable sigil ("*$var" looks unsightly).

Given the existence of spread, the (any-of:) macro is no longer needed.

Also: added a (range:) macro which, like its Python equivalent, creates an integer range array. I wanted to add range literal syntax, but I've decided I'm perhaps jumping the gun on literals at the moment, and should take it slow.