Replaced the sensor macros with the (live:) macro.
This macro, as I may've hinted before, simplifies the evaluation model of Twine 2 macros, eliminating the special case that sensors had. Rather than its expressions being stored in a thunk, they are now evaluated eagerly as normal, and now rely on the enclosing hook's (if:) macros to provide the conditional. In essence: (whenever: $g is true)[Bananas] becomes (live:20ms)[(if: $g is true)[Bananas]]. There is, of course, an increased verbosity to this syntax, but its simplified conceptual model, and composability with other macros, provides a greater boon.

(But, this macro isn't yet sufficient to replicate (when:), at least without clunky signalling temp variables. The next commit will add an (event:) macro which brings (when:)'s one-time behaviour to the scene.)

I've removed all occurrences of thunks in the code - now, all macros receive their arguments as bare values, which incidentally brings a slight performance gain.

Also:

 * Updated several test files' code.

 * Changed the (if:) macro so that its default transition is "instant" - this allows greater composition with (live:), and also fixes some transition quirks associated with it.