Replaced (event:) with a more general (stop:) command used within (live:).
After some reflection I've concluded that the distinction between (event:) and (live:) is a bit too high-level and ill-distinguished. I prefer having just a single "live" macro, (live:), rather than two with subtly different (and crumbily-defined) behaviour. So, I've fallen back to an alternative I had in mind: a (stop:) command which, when appearing inside a live hook, stops the nearest (live:) from repeating. This is more verbose, of course, and I don't quite like the look of it, but it's what is called for.

Also: fixed a bug where anonymous hooks couldn't appear just before HTML tags. (This caused certain macros, such as (background:), to break indirectly.)