Removed (cloak:), (uncloak:), (css:), (remove:), (nobr:)
(remove:) is one whose name I'm not keen on, and I've decided to remove it in place of the (replace:)[] idiom. (css:) is not a recommended use-case and can be accomplished with HTML, (nobr:) is fulfilled by the curly-brace syntax, and (cloak:) and (uncloak:) currently don't work too well with the rendering flow of sections. Maybe those will return in the future, but they're not too usable for 1.0.