Removed Section.updateEnchantments(), as it doesn't actually work as intended.
Section.updateEnchantments() was designed to keep all the enchantments in a section up-to-date even in the face of nebulous DOM manipulations to the passage. However, it turns out that few situations rely on it, and its presence caused some bugs where multiple enchantments refer to the same hook or text. So, I'm trying to go on without it.

Also fixed a bug where the header syntax didn't work on the first line of a passage.