Computed property access now allows number indices.
Formerly, supplying a number as a computed property index would fail unless you used (text:). This is now fixed, and in addition to that, you can now supply numbers as 1-indexed indices to arrays and strings. So, "Red"'s ($i), where $i is 1, would be the same as "Red"'s 1st. Note, however, that if $i was the string "1st", it would also work too - but not if it was just the string "1". Hmm.