Added "of" property indexing as a counterpart to possessive indexing.
Now, you can alternatively write "last of $a" instead of "$a's last", or "passages of $style" instead of "$style's passages". This is intended to provide a little more flexibility in phrasing/naming collection variables - although whether it succeeds is in question.

This syntax should also work with computed indexing ("(1 + 2) of $a") and it indexing ("1st of it").

Also, there's some instrumentation for adding a "page" identifier mixed in here.