Preliminary work on the (savegame:) and (loadgame:) macros.
These command macros should serve to implement a SugarCube-like localStorage save file ability. Currently, though, only some of the bare essentials are present, so they're not quite ready for public use.

To temporarily waylay the issue of serialising complex data structures, (savegame:) is currently specified to not work if an object has ever been saved to a Twine variable. This is, of course, somewhat dire a restriction, but I hope to slowly relax it in the future.

Also fixed some issues in the align syntax test code.