Fixed a bug where operators were incorrectly matched within unmatched words.
As an example, the bare text "whoishe" would be recognised as [text][is][text], even though in Patterns the "is" operator had a \b preceding its regex. As I should've known, the parsing method used by Lexer makes a leading \b meaningless. So, this restriction is now implemente with cannotFollow: ["text"] in these tokens' Markup rules, an adjustment to Lexer to allow "text" to be a legitimate cannotFollow.

The result: Javascript text snuck into expressions (like (print: navigator.userAgent)) would no longer be mangled if it contained a basic operator (like "or").

Also:

 * Whitespace is now internally recognised by TwineMarkup as distinct from text, and outputs a "whitespace" token instead of a "text" token for it.

 * Added tests for the "is in" operator (the reverse of "contains").

 * <script> tags in HTML may now execute during ChangeDescriptor.render(). However, their behaviour is not well-defined yet - it's unclear even to me what DOM they would have access to. This will have to be sorted out at some point.