Fixed some "it" bugs and returned the It variable to Operations.
Taking off from the previous commit, this makes it so that VarRef.create() no longer implicitly sets the It variable - rather, the new Operation.setIt() method must be wrapped around it, which the Compiler now does.

This also indirectly fixes an introduced bug with it, whereby (print: $red > $blue and it < $green) would interpret the second comparison as "$blue < green" instead of "$red < green".