Added the (align:) style changer macro.
(align:) is a macro form of the aligner syntax, and accepts a string containing an ASCII arrow of the same type that makes up the syntax ('==>', '=><==', etc).

Also, internally changed references to change descriptors' "code" to "prose", which I think is closer to the true content of that property. This also causes $Passages's passage datamaps to use 'prose' in place of 'code' now.