Improved the collapsing syntax's handling of whitespace considerably.
Now, all whitespace on the ends of the interior will be removed, and all runs between non-whitespace will be reduced to single spaces, regardless of the topology of the interior and what elements are contained within.

The collapsing syntax now creates a <tw-collapsed> container element, and its contents are collapsed at DOM insertion time, not parse time. This may seem inelegant, but it has a hidden boon: the collapsing syntax can now be made visible in debug mode. An easy trade.