Computed indexing has been converted to explicit indexing operators.
Now, you can use any expression in what was the computed index position, without needing to enclose it in brackets. This allows common types of indices, such as strings, to be more in-line with 'bareword' indices, instead of requiring both brackets and quotes. Considering that certain built-in datamaps like $Design may have indices like "Hovering Links", this is fairly significant.