Datamaps may now have numbers as data names.
You may now write (datamap: 1, "Red", 2, "Blue") without having to change the numbers to strings. Note that this may cause user confusion if they try to add or modify a data value whose name is number 1 using the string "1". Evidently a warning of some sort should be made when this occurs.

Also, the (datamap:) macro will now report errors when the wrong data type is in name position.