Changed the $Saves variable into a (saves:) macro.
I've decided to adopt the approach that read-only system variables should be exposed as getter-macros instead of variables, on the basis that their read-only nature is more obvious then. Of course, I'm still not fond of zero-arity macros, but I've come to accept that this is a minor trade-off for clarity in the variable interface (and, of course, less names squatted by system variables).