IFixed the escaped line syntax to work with headings and list items.
Now, you can split a heading using the escaped line syntax, and it will remain as a single list.

Also, decided to remove the heading syntax's tendency to remove trailing # marks, such as in "### HEADING ###", because of difficulty in getting it to parse, its lack of documentation, and its slight unintuitiveness compared to the bullet and list item syntax.