Added custom type signature error messages to (if:) and (else-if:)
These provide a hint to the author about how to correct the error. I deem this necessary because that addition of (if:)'s tighter type signature is a breaking change from 1.0.1, and this should provide a little migration help for those whose bugs were silently ignored until now.

Also added test specs for 'and', 'or' and 'not'.