Refined the (textstyle:"blur") fix to be more comprehensive.
If the user has complicated story CSS, it's not possible to determine what colour should be used for a blur hook via .css() until it's connected to the DOM. So, now the style application in ChangeDescriptor.update() is deferred via setTimeout(). This is of course a somewhat brittle fix, but a better solution isn't yet forthcoming.

Also fixed some bugs that broke certain type signatures in the previous commit.