Partially fixed the behaviour of (replace:) replacing out of collapsing regions.
Now, using (replace:) inside the collapsing syntax to change a hook outside of it, will correctly collapse the attached hook's contents as it's copied into the distant hook. But the reverse scenario (replacing into the collapsing syntax) still needs to be implemented.

This also fixes a bug where ChangeDescriptor.create() was overriding the parent descriptor's attr and styles arrays.