Finished fixing the (replace:) behaviour in the previous commit.
Now, using (replace:) from outside the collapsing syntax to alter a hook inside it will correctly use the original location's whitespace rules rather than the destination. This fix took some finagling, and it has no symmetry with the previous commit's fix, so you can understand I'm a little bothered by it. But, it seems to hold.