Split off an Enchantment internal type from the Enchantments macro module.
This internal type allows the concept of an "enchantment" to be extended beyond just the act of styling clickable links created by (click:) and such, and can be a mechanism by which general layout macros could be implemented. This also separates the act of styling the sections from the enchantment event function attached by the macros, which is specific to (click:)'s implementation.

Also, changed the alignment of some lines in the Patterns file.