Added a rudimentary implementation of HTML element selectors.
Now, enchantment and revision macros like (replace:), when given a string formatted like "<tag>", will select HTML tags in the passage, or the <tw-story> itself if the string is "<tw-story>". This allows such macros to remain scoped to their passages, while still allowing the <tw-story> to be enchanted. This is the basis for page-wide CSS-esque styling.

As this is a preliminary syntax enhancement, there's no test case for it yet.

Also:

 * Reinstated the reverted CSS changes from commit 274.

 * Improved (background:) to now set the enchantment's display to "block" if it probably contains a block element.