Made certain values (like AssignmentRequests) unable to be stored or used in macros.
It always seemed a bit suspicious that values like "$a to 1" or even the return value of (set:) could be used in expression position inside macros without warranting an error. So, now this is the case - AssignmentRequests must now only be in (set:) or (put:) and not also any macro that accepts Any, and (set:)/(put:) can't be used inside other macros, but only in passage prose.

Also added some tests related to direct variable printing and macro syntax.