Converted (if:)/(unless:)/(elseif:)/(else:) to be changer macros rather than boolean value macros.
This change, a TODO for awhile, provides the following benefits: 1) authors are warned when their (if:) isn't attached to a hook, 2) it allows (if:) to be composed with other changers, such as (text-style:). However, there is, I admit, a potential problem: On creating a composed changer using, say, (if: $poweredUp), changing the value of $poweredUp won't change the value of the composed changer. This can be seen as functional side-effect hygeine, but it may be puzzling to authors.

Also fixed the <audio> tests to work on browsers which lack Ogg support.