Fixed the (move:) macro, and improved its behaviour with hook references.
Now, the (move:) macro should again correctly delete the value from the source after copying it to the destination. Also, new semantics for deleting a hook reference have been implemented: (move: ?a to $a) will, after copying, replace ?a's content with the empty string.

Tests for the above have also been added.