Migrated build tools from Grunt to Make.
This toolshift should result in better running times for basic tasks, and a somewhat cleaner build script.

Also, as a consequence, proper jshintrc files have been created in each codebase (the Harlowe source, and the test specs). However, due to the loss of grunt-contrib-jshint, a custom jshint reporter adapted from it must be included to retain its desirable reporting messages.