Began developing a (goto-transition:) variant of (goto:)
This variant accepts, after the passage name, the name of a transition supported by (transition:). This allows the destination passage to use a specific transition animation. I'm still designing the syntax for this, so this will probably not be the final end-user design.

Also tweaked the transition SCSS to use variables.