Implemented the (transition-time:) style macro.
While a structure for this has existed in the changeDescriptor code for awhile, this commit finally instates the macro and transition-code conditions that uses it.

Note: this macro's command must be composed (added to) the (transition:) macro's command to be of use. It's speculative whether or not composing this with a command which doesn't affect the transition should cause an error or not. (Maybe only when attached to a hook?)