Changed lambda syntax to use ==>, and made params use an underscore sigil.
I feel that requiring a sigil for lambda params better distinguishes that their names are user-defined, and that they are essentially similar to proper variables. This also has the benefit of not requiring a specific keyword collision check.

Also, I decided that having a different character structure rather than "to" was necessary in order to implement different kinds of lambda macros, in particular ones approximating JS's every() and some().