Passage link syntax can now be used in expression position.
As these desugar to (link-goto:) calls in inline prose, it was decided that they might as well be present inside macros' calls as well. Square brackets had previously been reserved, so they can be safely employed here. (This will conflict with JS arrays, but only in the context of doubly-nested array literals, which are almost nonexistant.)

This also revealed that the ordering of rules in markup.js is more than a bit brittle, and should probably be reified in the near future.