Hook-attached macros may now have whitespace between them and their hooks.
This means that "(if: $x)  [text]", "(if: $x)\n[text]", and such are now syntactically acceptable - the whitespace is removed, and the macro is treated as if directly attached. Erring on the side of simplicity, any amount of whitespace and line breaks is allowed. Closes #3.

(But, a lingering issue: macros whose return values cannot be attached to hooks do not currently produce errors when they are.)