Anonymous hook syntax is now recognised without qualification.
You can now write "[text]", and the square brackets are, by themselves treated as a hook, albeit with no name (it cannot be referenced) and no attached changers. This means that all uses of matching single-square-brackets are considered to be hooks, regardless of context. This, I believe, slightly simplifies the syntax of the language and the cognitive load in understanding when square brackets will be displayed literally and when they will not.

Incidentally, this also changes the rules for when an expression attaches to a hook: now, commands will attach, but other kinds of data will not. The consequences for false negatives isn't very significant with this rule.