Reworked the lambda syntax, and added (folded:).
Now, the leading 'each' is gone, and furthermore, lambdas now consist of one or more of the following clauses, in any order: 'where', which filters value, 'via', which computes a new value, 'making', which specifies an additional variable as an accumulator (for use with reduce()-type macros), and 'with', which specifies an additional variable as a second parameter (for use with sort()-type macros). Additionally, due to the insolubility of the comma confusion issue (and the new 'making' and 'with' clauses), lambdas now only permit one leading parameter name.

I'm feeling much more confident about this syntax than before, but still have cold feet about 'via' and 'making'.

(folded:) is the name I've decided for the reduce() equivalent macro, winning out mainly because of its simple terseness (even if it isn't too clear about what its algorithm involves).