It is now an error to mutate collections that aren't in variables.
So, for instance, (set: (a:)'s 1st to 1) is now an error, because it accomplishes nothing, as is (set: (passage:)'s name to "X") etc. Using (set:) on hookrefs, however, still works for now.

This also fixes a few bugs related to the printing of computed property chains in error messages (such as "a datamap's ('E')").