Removed (click-page:), incorporating its functionality into (click: ?Page).
Now, including the ?Page, ?Passage or ?Sidebar in a (click:) enchantment's targets will, instead of turning its entire text into a link, border the page and wait for a click anywhere in it, like the (click-page:) macro provided.

You may observe that this "border-link" feature would be desirable for plain hooks. I'll have to see what I can do.