Lambdas may now omit the leading temp variable, and can use "it" for it.
I'm still not entirely sold on this decision, which is why it's currently undocumented, but here it is: you can now write (find: where it > 3) etc., and "it" will refer to the loop variable, even if elided. "it" also works when the variable is provided - (find: _item where it > 3).

(What I'm hesitant about is that this is ultimately a new meaning for "it", similar to but not equivalent to previous uses where "it"'s value was always leftward and explicitly visible.)

Also fixed/removed a few revision tests, and altered some documentation references to (datamap:), changing it to (dm:).