Added "any" and "all" properties to sequentials, which are used for "all" and "any" comparison operations.
These properties produce special, unstorable "determiner" values, which, when used as the object of "contains", "is in", or other comparison expressions, results in every element of the sequential being compared using the operation. This is, I feel, a better solution to simply adding "contains all" and "contains any" operators, in that it naturally expands to cover ">", ">=" and other non-set-related operations.

Also fixes a bug (that didn't manifest until just now) where certain unstorable values could be stored if they were in storable data structures.