Added the (hover-style:) changer macro, which allows passed style changers to only apply when hovering over the hook.
I'm a little uncomfortable about its name (as compared to (text-style:)) and the specificity of only allowing style changers (which currently does not include (transition:)s) but I think it'll be fine, and could be safely expanded a little in future.

Also fixed a bug where joining a changer, using "+" in markup, with a changer macro that produced an error, would suppress the error.