Added the "each" lambda keyword, to enable no-clause lambdas to be expressed.
The "each" keyword is a lambda "clause" that simply sets the lambda's subject, but, importantly, it allows one to create a lambda with no actual clauses. Formerly, such a lambda would be syntactically identical to (and unable to be disambiguated from) a plain temp variable. This allows future macros (such as a possible (loop:) macro) to use lambdas that iterate over the entire sequence, without specifying any specific selection or modification.