Fixed (enchant: ?Page) not being able to override default CSS (or other CSS) on <tw-story>.
This fix forcibly checks which CSS properties are being applied to the <tw-enchantment> by style changers, and adds matching "inherit" inline styles to the <tw-story> below. Yes, it's unfortunate that the <tw-story> has to be permuted for this? oh well.

Also added CodeMirror mode CSS that keeps the line-break marker gutter at left:0, which, for reasons yet unknown, it sometimes isn't.

Version changed to 2.0.1.

Closes #85. Closes #83.