Changed the ~~ markup to produce a strikethrough style instead of a censor-bar style.
The censor-bar style, which was used in all previous versions but not correctly documented, was bugged to always be black even if the text colour was not black. It can be replicated in stories by simply using a `(background-colour:)` macro (preferably set to a variable) in its place.

Also re-fixed certain broken maths macros for probably the fourth time in Harlowe history.

Also reverted the default line-height CSS for <h1> and other elements to "1em" instead of "1rem", due to it causing problems with line-wrapped headers.

Closes #115. Closes #107. Closes #80.